ft_transcendence/backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==3.2.13
django-allauth==0.63.1
django-cors-headers==3.7.0
django-redis==5.4.0
django_csp==3.8
djangorestframework==3.15.1
docker==7.0.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.4.0
pyOpenSSL==24.1.0
PySocks==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.2
selenium==4.21.0
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
websocket==0.2.1
websocket-client==1.8.0
websockets==10.1
whitenoise==6.6.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4